+++ spoon-runner/src/main/java/com/squareup/spoon/SpoonDeviceRunner.java
@@ -5,7 +5,6 @@ import com.android.ddmlib.CollectingOutputReceiver;
 import com.android.ddmlib.DdmPreferences;
 import com.android.ddmlib.IDevice;
 import com.android.ddmlib.InstallException;
-import com.android.ddmlib.SyncService;
 import com.android.ddmlib.logcat.LogCatMessage;
 import com.android.ddmlib.testrunner.IRemoteAndroidTestRunner;
 import com.android.ddmlib.testrunner.ITestRunListener;
@@ -21,6 +20,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -28,9 +28,10 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.filefilter.TrueFileFilter;
 
 import static com.android.ddmlib.FileListingService.FileEntry;
+import static com.android.ddmlib.SyncService.getNullProgressMonitor;
 import static com.google.common.base.Strings.isNullOrEmpty;
-import static com.squareup.spoon.Spoon.SPOON_SCREENSHOTS;
 import static com.squareup.spoon.Spoon.SPOON_FILES;
+import static com.squareup.spoon.Spoon.SPOON_SCREENSHOTS;
 import static com.squareup.spoon.SpoonLogger.logDebug;
 import static com.squareup.spoon.SpoonLogger.logError;
 import static com.squareup.spoon.SpoonLogger.logInfo;
@@ -45,16 +46,20 @@ public final class SpoonDeviceRunner {
   private static final String FILE_RESULT = "result.json";
   private static final String DEVICE_SCREENSHOT_DIR = "app_" + SPOON_SCREENSHOTS;
   private static final String DEVICE_FILE_DIR = "app_" + SPOON_FILES;
-  private static final String [] DEVICE_DIRS = {DEVICE_SCREENSHOT_DIR, DEVICE_FILE_DIR};
+  private static final String[] DEVICE_DIRS = {DEVICE_SCREENSHOT_DIR, DEVICE_FILE_DIR};
   static final String TEMP_DIR = "work";
   static final String JUNIT_DIR = "junit-reports";
   static final String IMAGE_DIR = "image";
   static final String FILE_DIR = "file";
+  static final String COVERAGE_FILE = "coverage.ec";
+  static final String COVERAGE_DIR = "coverage";
 
   private final File sdk;
   private final File apk;
   private final File testApk;
   private final String serial;
+  private final int shardIndex;
+  private final int numShards;
   private final boolean debug;
   private final boolean noAnimations;
   private final int adbTimeout;
@@ -65,10 +70,13 @@ public final class SpoonDeviceRunner {
   private final File work;
   private final File junitReport;
   private final File imageDir;
+  private final File coverageDir;
   private final File fileDir;
   private final String classpath;
   private final SpoonInstrumentationInfo instrumentationInfo;
+  private boolean codeCoverage;
   private final List<ITestRunListener> testRunListeners;
+  private final boolean grantAll;
 
   /**
    * Create a test runner for a single device.
@@ -84,18 +92,20 @@ public final class SpoonDeviceRunner {
    * @param instrumentationInfo Test apk manifest information.
    * @param className Test class name to run or {@code null} to run all tests.
    * @param methodName Test method name to run or {@code null} to run all tests.  Must also pass
-   *        {@code className}.
+   * {@code className}.
    * @param testRunListeners Additional TestRunListener or empty list.
    */
-  SpoonDeviceRunner(File sdk, File apk, File testApk, File output, String serial, boolean debug,
-      boolean noAnimations, int adbTimeout, String classpath,
+  SpoonDeviceRunner(File sdk, File apk, File testApk, File output, String serial, int shardIndex,
+      int numShards, boolean debug, boolean noAnimations, int adbTimeout, String classpath,
       SpoonInstrumentationInfo instrumentationInfo, List<String> instrumentationArgs,
       String className, String methodName, IRemoteAndroidTestRunner.TestSize testSize,
-      List<ITestRunListener> testRunListeners) {
+      List<ITestRunListener> testRunListeners, boolean codeCoverage, boolean grantAll) {
     this.sdk = sdk;
     this.apk = apk;
     this.testApk = testApk;
     this.serial = serial;
+    this.shardIndex = shardIndex;
+    this.numShards = numShards;
     this.debug = debug;
     this.noAnimations = noAnimations;
     this.adbTimeout = adbTimeout;
@@ -105,13 +115,15 @@ public final class SpoonDeviceRunner {
     this.testSize = testSize;
     this.classpath = classpath;
     this.instrumentationInfo = instrumentationInfo;
-
+    this.codeCoverage = codeCoverage;
     serial = SpoonUtils.sanitizeSerial(serial);
     this.work = FileUtils.getFile(output, TEMP_DIR, serial);
     this.junitReport = FileUtils.getFile(output, JUNIT_DIR, serial + ".xml");
     this.imageDir = FileUtils.getFile(output, IMAGE_DIR, serial);
     this.fileDir = FileUtils.getFile(output, FILE_DIR, serial);
+    this.coverageDir = FileUtils.getFile(output, COVERAGE_DIR, serial);
     this.testRunListeners = testRunListeners;
+    this.grantAll = grantAll;
   }
 
   /** Serialize to disk and start {@link #main(String...)} in another process. */
@@ -178,35 +190,43 @@ public final class SpoonDeviceRunner {
 
     // Now install the main application and the instrumentation application.
     try {
-      device.installPackage(apk.getAbsolutePath(), true);
+      String extraArgument = "";
+      if (grantAll && deviceDetails.getApiLevel() >= DeviceDetails.MARSHMALLOW_API_LEVEL) {
+        extraArgument = "-g";
+      }
+      device.installPackage(apk.getAbsolutePath(), true, extraArgument);
     } catch (InstallException e) {
       logInfo("InstallException while install app apk on device [%s]", serial);
       e.printStackTrace(System.out);
-      return result.markInstallAsFailed("Unable to install application APK.").build();
+      return result.markInstallAsFailed(
+              "Unable to install application APK.").addException(e).build();
     }
     try {
       device.installPackage(testApk.getAbsolutePath(), true);
     } catch (InstallException e) {
       logInfo("InstallException while install test apk on device [%s]", serial);
       e.printStackTrace(System.out);
-      return result.markInstallAsFailed("Unable to install instrumentation APK.").build();
+      return result.markInstallAsFailed(
+              "Unable to install instrumentation APK.").addException(e).build();
     }
 
     // If this is Android Marshmallow or above grant WRITE_EXTERNAL_STORAGE
-    if (Integer.parseInt(device.getProperty(IDevice.PROP_BUILD_API_LEVEL)) >= 23) {
+    if (deviceDetails.getApiLevel() >= DeviceDetails.MARSHMALLOW_API_LEVEL) {
       String appPackage = instrumentationInfo.getApplicationPackage();
       try {
         CollectingOutputReceiver grantOutputReceiver = new CollectingOutputReceiver();
-        device.executeShellCommand("pm grant " + appPackage
-            + " android.permission.READ_EXTERNAL_STORAGE", grantOutputReceiver);
-        device.executeShellCommand("pm grant " + appPackage
-            + " android.permission.WRITE_EXTERNAL_STORAGE", grantOutputReceiver);
+        device.executeShellCommand(
+            "pm grant " + appPackage + " android.permission.READ_EXTERNAL_STORAGE",
+            grantOutputReceiver);
+        device.executeShellCommand(
+            "pm grant " + appPackage + " android.permission.WRITE_EXTERNAL_STORAGE",
+            grantOutputReceiver);
       } catch (Exception e) {
         logInfo("Exception while granting external storage access to application apk"
             + "on device [%s]", serial);
         e.printStackTrace(System.out);
-        return result.markInstallAsFailed("Unable to grant external storage access to"
-            + " application APK.").build();
+        return result.markInstallAsFailed(
+            "Unable to grant external storage access to application APK.").addException(e).build();
       }
     }
 
@@ -224,12 +244,28 @@ public final class SpoonDeviceRunner {
 
       if (instrumentationArgs != null && instrumentationArgs.size() > 0) {
         for (String pair : instrumentationArgs) {
-          String[] kvp = pair.split("=");
-          if (kvp.length != 2 || isNullOrEmpty(kvp[0]) || isNullOrEmpty(kvp[1])) {
+          int firstEqualSignIndex = pair.indexOf("=");
+          if (firstEqualSignIndex <= -1) {
+            // No Equal Sign, can't process
+            logDebug(debug, "Can't process instrumentationArg [%s] (no equal sign)", pair);
             continue;
           }
-          runner.addInstrumentationArg(kvp[0], kvp[1]);
+          String key = pair.substring(0, firstEqualSignIndex);
+          String value = pair.substring(firstEqualSignIndex + 1);
+          if (isNullOrEmpty(key) || isNullOrEmpty(value)) {
+            // Invalid values, skipping
+            logDebug(debug, "Can't process instrumentationArg [%s] (empty key or value)", pair);
+            continue;
+          }
+          runner.addInstrumentationArg(key, value);
         }
+      }
+      if (codeCoverage) {
+        addCodeCoverageInstrumentationArgs(runner, device);
+      }
+      // Add the sharding instrumentation arguments if necessary
+      if (numShards != 0) {
+        addShardingInstrumentationArgs(runner);
       }
 
       if (!isNullOrEmpty(className)) {
@@ -258,26 +294,61 @@ public final class SpoonDeviceRunner {
     try {
       logDebug(debug, "About to grab screenshots and prepare output for [%s]", serial);
       pullDeviceFiles(device);
-
-      File screenshotDir = new File(work, DEVICE_SCREENSHOT_DIR);
-      File testFilesDir = new File(work, DEVICE_FILE_DIR);
-      if (screenshotDir.exists()) {
-        imageDir.mkdirs();
-
-        handleImages(result, screenshotDir);
-        FileUtils.deleteDirectory(screenshotDir);
+      if (codeCoverage) {
+        pullCoverageFile(device);
       }
-      if (testFilesDir.exists()) {
-        fileDir.mkdirs();
-        handleFiles(result, testFilesDir);
-        FileUtils.deleteDirectory(testFilesDir);
-      }
+
+      cleanScreenshotsDirectory(result);
+      cleanFilesDirectory(result);
+
     } catch (Exception e) {
       result.addException(e);
     }
     logDebug(debug, "Done running for [%s]", serial);
 
     return result.build();
+  }
+
+  private void addCodeCoverageInstrumentationArgs(RemoteAndroidTestRunner runner, IDevice device)
+          throws Exception {
+    String coveragePath = getExternalStoragePath(device, COVERAGE_FILE);
+    runner.addInstrumentationArg("coverage", "true");
+    runner.addInstrumentationArg("coverageFile", coveragePath);
+  }
+
+  private void addShardingInstrumentationArgs(RemoteAndroidTestRunner runner) {
+    runner.addInstrumentationArg("numShards", Integer.toString(numShards));
+    runner.addInstrumentationArg("shardIndex", Integer.toString(shardIndex));
+  }
+
+  private void cleanScreenshotsDirectory(DeviceResult.Builder result) throws IOException {
+    File screenshotDir = new File(work, DEVICE_SCREENSHOT_DIR);
+    if (screenshotDir.exists()) {
+      imageDir.mkdirs();
+      handleImages(result, screenshotDir);
+      FileUtils.deleteDirectory(screenshotDir);
+    }
+  }
+
+  private void cleanFilesDirectory(DeviceResult.Builder result) throws IOException {
+    File testFilesDir = new File(work, DEVICE_FILE_DIR);
+    if (testFilesDir.exists()) {
+      fileDir.mkdirs();
+      handleFiles(result, testFilesDir);
+      FileUtils.deleteDirectory(testFilesDir);
+    }
+  }
+
+  private void pullCoverageFile(IDevice device) {
+    coverageDir.mkdirs();
+    File coverageFile = new File(coverageDir, COVERAGE_FILE);
+    String remotePath;
+    try {
+      remotePath = getExternalStoragePath(device, COVERAGE_FILE);
+    } catch (Exception exception) {
+      throw new RuntimeException("error while calculating coverage file path.", exception);
+    }
+    adbPullFile(device, remotePath, coverageFile.getAbsolutePath());
   }
 
   private void handleImages(DeviceResult.Builder result, File screenshotDir) throws IOException {
@@ -332,8 +403,7 @@ public final class SpoonDeviceRunner {
   private void handleFiles(DeviceResult.Builder result, File testFileDir) throws IOException {
     File[] classNameDirs = testFileDir.listFiles();
     if (classNameDirs != null) {
-      logInfo("Found class name dirs: " + classNameDirs);
-      Multimap<DeviceTest, File> testFiles = ArrayListMultimap.create();
+      logInfo("Found class name dirs: " + Arrays.toString(classNameDirs));
       for (File classNameDir : classNameDirs) {
         String className = classNameDir.getName();
         File destDir = new File(fileDir, className);
@@ -342,7 +412,7 @@ public final class SpoonDeviceRunner {
 
         // Get a sorted list of all files from the device run.
         List<File> files = new ArrayList<File>(
-                FileUtils.listFiles(destDir, TrueFileFilter.INSTANCE, TrueFileFilter.INSTANCE));
+            FileUtils.listFiles(destDir, TrueFileFilter.INSTANCE, TrueFileFilter.INSTANCE));
         Collections.sort(files);
 
         // Iterate over each file and associate it with its
@@ -350,8 +420,8 @@ public final class SpoonDeviceRunner {
         for (File file : files) {
           String methodName = file.getParentFile().getName();
           DeviceTest testIdentifier = new DeviceTest(className, methodName);
-          final DeviceTestResult.Builder resultBuilder
-                  = result.getMethodResultBuilder(testIdentifier);
+          final DeviceTestResult.Builder resultBuilder =
+              result.getMethodResultBuilder(testIdentifier);
           if (resultBuilder != null) {
             resultBuilder.addFile(file);
             logInfo("Added file as result: " + file + " for " + testIdentifier);
@@ -372,11 +442,11 @@ public final class SpoonDeviceRunner {
 
   private void pullDirectory(final IDevice device, final String name) throws Exception {
     // Output path on private internal storage, for KitKat and below.
-    FileEntry internalDir = getScreenshotDirOnInternalStorage(name);
+    FileEntry internalDir = getDirectoryOnInternalStorage(name);
     logDebug(debug, "Internal path is " + internalDir.getFullPath());
 
     // Output path on public external storage, for Lollipop and above.
-    FileEntry externalDir = getScreenshotDirOnExternalStorage(device, name);
+    FileEntry externalDir = getDirectoryOnExternalStorage(device, name);
     logDebug(debug, "External path is " + externalDir.getFullPath());
 
     // Sync test output files to the local filesystem.
@@ -390,30 +460,42 @@ public final class SpoonDeviceRunner {
 
   private void adbPull(IDevice device, FileEntry remoteDirName, String localDirName) {
     try {
-      device.getSyncService()
-          .pull(new FileEntry[] {remoteDirName}, localDirName,
-              SyncService.getNullProgressMonitor());
+      device.getSyncService().pull(new FileEntry[]{remoteDirName}, localDirName,
+          getNullProgressMonitor());
     } catch (Exception e) {
       logDebug(debug, e.getMessage(), e);
     }
   }
 
-  private FileEntry getScreenshotDirOnInternalStorage(final String dir) {
-    String appPackage = instrumentationInfo.getApplicationPackage();
-    String internalPath = "/data/data/" + appPackage + "/" + dir;
+  private void adbPullFile(IDevice device, String remoteFile, String localDir) {
+    try {
+      device.getSyncService()
+          .pullFile(remoteFile, localDir, getNullProgressMonitor());
+    } catch (Exception e) {
+      logDebug(debug, e.getMessage(), e);
+    }
+  }
+
+  private FileEntry getDirectoryOnInternalStorage(final String dir) {
+    String internalPath = getInternalPath(dir);
     return obtainDirectoryFileEntry(internalPath);
   }
 
-  private static FileEntry getScreenshotDirOnExternalStorage(IDevice device, final String dir)
-          throws Exception {
-    String externalPath = getExternalStoragePath(device) + "/" + dir;
+  private String getInternalPath(String path) {
+    String appPackage = instrumentationInfo.getApplicationPackage();
+    return "/data/data/" + appPackage + "/" + path;
+  }
+
+  private FileEntry getDirectoryOnExternalStorage(IDevice device, final String dir)
+      throws Exception {
+    String externalPath = getExternalStoragePath(device, dir);
     return obtainDirectoryFileEntry(externalPath);
   }
 
-  private static String getExternalStoragePath(IDevice device) throws Exception {
+  private String getExternalStoragePath(IDevice device, final String path) throws Exception {
     CollectingOutputReceiver pathNameOutputReceiver = new CollectingOutputReceiver();
     device.executeShellCommand("echo $EXTERNAL_STORAGE", pathNameOutputReceiver);
-    return pathNameOutputReceiver.getOutput().trim();
+    return pathNameOutputReceiver.getOutput().trim() + "/" + path;
   }
 
   /** Grab all the parsed logs and map them to individual tests. */