+++ spoon-runner/src/main/java/com/squareup/spoon/SpoonRunner.java
@@ -7,9 +7,9 @@ import com.beust.jcommander.IStringConverter;
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.ParameterException;
+import com.beust.jcommander.converters.IParameterSplitter;
 import com.google.common.collect.ImmutableSet;
 import com.squareup.spoon.html.HtmlRenderer;
-
 import org.apache.commons.io.FileUtils;
 
 import java.io.BufferedReader;
@@ -39,7 +39,7 @@ import static java.util.Collections.synchronizedSet;
 public final class SpoonRunner {
   private static final String DEFAULT_TITLE = "Spoon Execution";
   public static final String DEFAULT_OUTPUT_DIRECTORY = "spoon-output";
-  private static final int DEFAULT_ADB_TIMEOUT = 10 * 60; //10 minutes
+  private static final int DEFAULT_ADB_TIMEOUT_SEC = 10 * 60; //10 minutes
   private final ExecutorService threadExecutor;
 
   private final String title;
@@ -49,24 +49,29 @@ public final class SpoonRunner {
   private final File output;
   private final boolean debug;
   private final boolean noAnimations;
-  private final int adbTimeout;
+  private final int adbTimeoutMillis;
   private final List<String> instrumentationArgs;
   private final String className;
   private final String methodName;
   private final Set<String> serials;
+  private final Set<String> skipDevices;
+  private final boolean shard;
   private final String classpath;
   private final IRemoteAndroidTestRunner.TestSize testSize;
+  private boolean codeCoverage;
   private final boolean failIfNoDeviceConnected;
   private final List<ITestRunListener> testRunListeners;
   private final boolean terminateAdb;
   private File initScript;
+  private final boolean grantAll;
 
   private SpoonRunner(String title, File androidSdk, File applicationApk, File instrumentationApk,
-      File output, boolean debug, boolean noAnimations, int adbTimeout, Set<String> serials,
-      String classpath, List<String> instrumentationArgs, String className, String methodName,
-      IRemoteAndroidTestRunner.TestSize testSize, boolean failIfNoDeviceConnected,
-      List<ITestRunListener> testRunListeners, boolean sequential, File initScript,
-      boolean terminateAdb) {
+      File output, boolean debug, boolean noAnimations, int adbTimeoutMillis, Set<String> serials,
+      Set<String> skipDevices,
+      boolean shard, String classpath, List<String> instrumentationArgs, String className,
+      String methodName, IRemoteAndroidTestRunner.TestSize testSize,
+      boolean failIfNoDeviceConnected, List<ITestRunListener> testRunListeners, boolean sequential,
+      File initScript, boolean grantAll, boolean terminateAdb, boolean codeCoverage) {
     this.title = title;
     this.androidSdk = androidSdk;
     this.applicationApk = applicationApk;
@@ -74,17 +79,21 @@ public final class SpoonRunner {
     this.output = output;
     this.debug = debug;
     this.noAnimations = noAnimations;
-    this.adbTimeout = adbTimeout;
+    this.adbTimeoutMillis = adbTimeoutMillis;
     this.instrumentationArgs = instrumentationArgs;
     this.className = className;
     this.methodName = methodName;
     this.classpath = classpath;
     this.testSize = testSize;
+    this.skipDevices = skipDevices;
+    this.codeCoverage = codeCoverage;
     this.serials = ImmutableSet.copyOf(serials);
+    this.shard = shard;
     this.failIfNoDeviceConnected = failIfNoDeviceConnected;
     this.testRunListeners = testRunListeners;
     this.terminateAdb = terminateAdb;
     this.initScript = initScript;
+    this.grantAll = grantAll;
 
     if (sequential) {
       this.threadExecutor = Executors.newSingleThreadExecutor();
@@ -102,20 +111,25 @@ public final class SpoonRunner {
     checkArgument(applicationApk.exists(), "Could not find application APK.");
     checkArgument(instrumentationApk.exists(), "Could not find instrumentation APK.");
 
-    AndroidDebugBridge adb = SpoonUtils.initAdb(androidSdk, adbTimeout);
+    AndroidDebugBridge adb = SpoonUtils.initAdb(androidSdk, adbTimeoutMillis);
 
     try {
+      final SpoonInstrumentationInfo testInfo = parseFromFile(instrumentationApk);
+
       // If we were given an empty serial set, load all available devices.
       Set<String> serials = this.serials;
       if (serials.isEmpty()) {
-        serials = SpoonUtils.findAllDevices(adb);
+        serials = SpoonUtils.findAllDevices(adb, testInfo.getMinSdkVersion());
+      }
+      if (this.skipDevices != null && !this.skipDevices.isEmpty()) {
+        serials.removeAll(this.skipDevices);
       }
       if (failIfNoDeviceConnected && serials.isEmpty()) {
         throw new RuntimeException("No device(s) found.");
       }
 
       // Execute all the things...
-      SpoonSummary summary = runTests(adb, serials);
+      SpoonSummary summary = runTests(adb, serials, testInfo);
       // ...and render to HTML
       new HtmlRenderer(summary, SpoonUtils.GSON, output).render();
 
@@ -127,7 +141,8 @@ public final class SpoonRunner {
     }
   }
 
-  private SpoonSummary runTests(AndroidDebugBridge adb, Set<String> serials) {
+  private SpoonSummary runTests(AndroidDebugBridge adb, Set<String> serials,
+      final SpoonInstrumentationInfo testInfo) {
     int targetCount = serials.size();
     logInfo("Executing instrumentation suite on %d device(s).", targetCount);
 
@@ -137,7 +152,6 @@ public final class SpoonRunner {
       throw new RuntimeException("Unable to clean output directory: " + output, e);
     }
 
-    final SpoonInstrumentationInfo testInfo = parseFromFile(instrumentationApk);
     logDebug(debug, "Application: %s from %s", testInfo.getApplicationPackage(),
         applicationApk.getAbsolutePath());
     logDebug(debug, "Instrumentation: %s from %s", testInfo.getInstrumentationPackage(),
@@ -156,7 +170,7 @@ public final class SpoonRunner {
       String safeSerial = SpoonUtils.sanitizeSerial(serial);
       try {
         logDebug(debug, "[%s] Starting execution.", serial);
-        summary.addResult(safeSerial, getTestRunner(serial, testInfo).run(adb));
+        summary.addResult(safeSerial, getTestRunner(serial, 0, 0, testInfo).run(adb));
       } catch (Exception e) {
         logDebug(debug, "[%s] Execution exception!", serial);
         e.printStackTrace(System.out);
@@ -171,13 +185,18 @@ public final class SpoonRunner {
       // Spawn a new thread for each device and wait for them all to finish.
       final CountDownLatch done = new CountDownLatch(targetCount);
       final Set<String> remaining = synchronizedSet(new HashSet<String>(serials));
+
+      int shardIndex = 0;
+      final int numShards = shard ? serials.size() : 0;
       for (final String serial : serials) {
         final String safeSerial = SpoonUtils.sanitizeSerial(serial);
         logDebug(debug, "[%s] Starting execution.", serial);
+        final int safeShardIndex = shardIndex;
         Runnable runnable = new Runnable() {
           @Override public void run() {
             try {
-              summary.addResult(safeSerial, getTestRunner(serial, testInfo).runInNewProcess());
+              summary.addResult(safeSerial,
+                  getTestRunner(serial, safeShardIndex, numShards, testInfo).runInNewProcess());
             } catch (Exception e) {
               e.printStackTrace(System.out);
               summary.addResult(safeSerial, new DeviceResult.Builder().addException(e).build());
@@ -189,7 +208,10 @@ public final class SpoonRunner {
             }
           }
         };
-
+        if (shard) {
+          shardIndex++;
+          logDebug(debug, "shardIndex [%d]", shardIndex);
+        }
         threadExecutor.execute(runnable);
       }
 
@@ -201,6 +223,16 @@ public final class SpoonRunner {
       }
     }
 
+    if (codeCoverage) {
+      SpoonCoverageMerger coverageMerger = new SpoonCoverageMerger();
+      try {
+        coverageMerger.mergeCoverageFiles(serials, output);
+        logDebug(debug, "Merging of coverage files done.");
+      } catch (IOException exception) {
+        throw new RuntimeException("Error while merging coverage files. "
+            + "Did you set the \"testCoverageEnabled\" flag in your build.gradle?", exception);
+      }
+    }
     if (!debug) {
       // Clean up anything in the work directory.
       try {
@@ -215,8 +247,7 @@ public final class SpoonRunner {
   /** Returns a {@link Runnable} to launch the script before/between devices in sequential mode. */
   private Runnable getRunnableScript() {
     return new Runnable() {
-      @Override
-      public void run() {
+      @Override public void run() {
         executeInitScript();
       }
     };
@@ -228,7 +259,7 @@ public final class SpoonRunner {
       try {
         Runtime run = Runtime.getRuntime();
         Process proc = run.exec(new String[] {
-                "/bin/bash", "-c", initScript.getAbsolutePath()
+            "/bin/bash", "-c", initScript.getAbsolutePath()
         });
         BufferedReader br = new BufferedReader(new InputStreamReader(proc.getInputStream()));
         String line;
@@ -255,8 +286,8 @@ public final class SpoonRunner {
       if (result.getInstallFailed()) {
         return false; // App and/or test installation failed.
       }
-      if (!result.getExceptions().isEmpty() && result.getTestResults().isEmpty()) {
-        return false; // No tests run and top-level exception present.
+      if (!result.getExceptions().isEmpty() || result.getTestResults().isEmpty()) {
+        return false; // Top-level exception present, or no tests were run.
       }
       for (DeviceTestResult methodResult : result.getTestResults().values()) {
         if (methodResult.getStatus() != Status.PASS) {
@@ -267,10 +298,12 @@ public final class SpoonRunner {
     return true;
   }
 
-  private SpoonDeviceRunner getTestRunner(String serial, SpoonInstrumentationInfo testInfo) {
+  private SpoonDeviceRunner getTestRunner(String serial, int shardIndex, int numShards,
+      SpoonInstrumentationInfo testInfo) {
     return new SpoonDeviceRunner(androidSdk, applicationApk, instrumentationApk, output, serial,
-        debug, noAnimations, adbTimeout, classpath, testInfo, instrumentationArgs, className,
-        methodName, testSize, testRunListeners);
+        shardIndex, numShards, debug, noAnimations, adbTimeoutMillis, classpath, testInfo,
+        instrumentationArgs, className, methodName, testSize, testRunListeners, codeCoverage,
+        grantAll);
   }
 
   /** Build a test suite for the specified devices and configuration. */
@@ -282,18 +315,22 @@ public final class SpoonRunner {
     private File output;
     private boolean debug = false;
     private Set<String> serials;
+    private Set<String> skipDevices;
     private String classpath = System.getProperty("java.class.path");
     private List<String> instrumentationArgs;
     private String className;
     private String methodName;
     private boolean noAnimations;
     private IRemoteAndroidTestRunner.TestSize testSize;
-    private int adbTimeout;
+    private int adbTimeoutMillis = DEFAULT_ADB_TIMEOUT_SEC * 1000;
     private boolean failIfNoDeviceConnected;
     private List<ITestRunListener> testRunListeners = new ArrayList<ITestRunListener>();
     private boolean sequential;
     private File initScript;
+    private boolean grantAll;
     private boolean terminateAdb = true;
+    private boolean codeCoverage;
+    private boolean shard = false;
 
     /** Identifying title for this execution. */
     public Builder setTitle(String title) {
@@ -347,7 +384,7 @@ public final class SpoonRunner {
 
     /** Set ADB timeout. */
     public Builder setAdbTimeout(int value) {
-      this.adbTimeout = value;
+      this.adbTimeoutMillis = value;
       return this;
     }
 
@@ -362,10 +399,23 @@ public final class SpoonRunner {
       return this;
     }
 
+    /** Add a device serial for skipping test execution. */
+    public Builder skipDevice(String serial) {
+      checkNotNull(serial, "Serial cannot be null.");
+      if (skipDevices == null) {
+        skipDevices = new LinkedHashSet<String>();
+      }
+      skipDevices.add(serial);
+      return this;
+    }
+
     /** Use all currently attached device serials when executed. */
     public Builder useAllAttachedDevices() {
       if (this.serials != null) {
         throw new IllegalStateException("Serial list already contains entries.");
+      }
+      if (this.skipDevices != null) {
+        logInfo("Skipping devices not empty, will skip following serials: ", skipDevices);
       }
       if (this.androidSdk == null) {
         throw new IllegalStateException("SDK must be set before calling this method.");
@@ -415,8 +465,23 @@ public final class SpoonRunner {
       return this;
     }
 
+    public Builder setGrantAll(boolean grantAll) {
+      this.grantAll = grantAll;
+      return this;
+    }
+
     public Builder setMethodName(String methodName) {
       this.methodName = methodName;
+      return this;
+    }
+
+    public Builder setCodeCoverage(boolean codeCoverage) {
+      this.codeCoverage = codeCoverage;
+      return this;
+    }
+
+    public Builder setShard(boolean shard) {
+      this.shard = shard;
       return this;
     }
 
@@ -444,9 +509,20 @@ public final class SpoonRunner {
       }
 
       return new SpoonRunner(title, androidSdk, applicationApk, instrumentationApk, output, debug,
-          noAnimations, adbTimeout, serials, classpath, instrumentationArgs, className, methodName,
-          testSize, failIfNoDeviceConnected, testRunListeners, sequential, initScript,
-          terminateAdb);
+          noAnimations, adbTimeoutMillis, serials, skipDevices, shard, classpath,
+          instrumentationArgs, className, methodName, testSize, failIfNoDeviceConnected,
+          testRunListeners, sequential, initScript, grantAll, terminateAdb, codeCoverage);
+    }
+  }
+
+  /**
+   * Instances where string input should not be split there needs to
+   * be a way override the default splitting behavior.
+   * This class provides that options
+   */
+  public static class NoSplitter implements IParameterSplitter {
+    @Override public List<String> split(String value) {
+      return Collections.singletonList(value);
     }
   }
 
@@ -462,9 +538,9 @@ public final class SpoonRunner {
         converter = FileConverter.class, required = true) //
     public File testApk;
 
-    @Parameter(names = { "--e" },
+    @Parameter(names = { "--e" }, variableArity = true, splitter = NoSplitter.class,
         description = "Arguments to pass to the Instrumentation Runner. This can be used multiple"
-            + " times for multiple entries. Usage: --e <NAME>=<VALUE>.")
+            + " times for multiple entries. Usage: --e <NAME>=<VALUE>.") //
     public List<String> instrumentationArgs;
 
     @Parameter(names = { "--class-name" }, description = "Test class name to run (fully-qualified)")
@@ -497,24 +573,40 @@ public final class SpoonRunner {
     public boolean sequential;
 
     @Parameter(names = { "--init-script" },
-            description = "Script file executed between each devices",
-            converter = FileConverter.class)
+        description = "Script file executed between each devices",
+        converter = FileConverter.class) //
     public File initScript;
+
+    @Parameter(names = { "--grant-all" },
+        description = "Grant all runtime permissions during installation on Marshmallow and above"
+                + " devices") //
+    public boolean grantAll;
 
     @Parameter(names = { "--no-animations" }, description = "Disable animated gif generation")
     public boolean noAnimations;
 
     @Parameter(names = { "--adb-timeout" },
         description = "Set maximum execution time per test in seconds (10min default)") //
-    public int adbTimeoutSeconds = DEFAULT_ADB_TIMEOUT;
+    public int adbTimeoutSeconds = DEFAULT_ADB_TIMEOUT_SEC;
 
     @SuppressWarnings("MismatchedQueryAndUpdateOfCollection") //
     @Parameter(names = "-serial",
         description = "Serial of the device to use (May be used multiple times)")
     private List<String> serials = new ArrayList<String>();
 
+    @Parameter(names = "-skipDevices",
+        description = "Serial of the device to skip (May be used multiple times)")
+    private List<String> skipDevices = new ArrayList<String>();
+
+    @Parameter(names = { "--shard" },
+        description = "Automatically shard across all specified serials") //
+    public boolean shard;
+
     @Parameter(names = { "--debug" }, hidden = true) //
     public boolean debug;
+
+    @Parameter(names = { "--coverage" }, description = "Code coverage flag", arity = 1)
+    public Boolean codeCoverage = false;
 
     @Parameter(names = { "-h", "--help" }, description = "Command help", help = true, hidden = true)
     public boolean help;
@@ -576,15 +668,25 @@ public final class SpoonRunner {
         .setFailIfNoDeviceConnected(parsedArgs.failIfNoDeviceConnected)
         .setSequential(parsedArgs.sequential)
         .setInitScript(parsedArgs.initScript)
+        .setGrantAll(parsedArgs.grantAll)
         .setInstrumentationArgs(parsedArgs.instrumentationArgs)
+        .setCodeCoverage(parsedArgs.codeCoverage)
         .setClassName(parsedArgs.className)
-        .setMethodName(parsedArgs.methodName);
+        .setMethodName(parsedArgs.methodName)
+        .setShard(parsedArgs.shard);
 
     if (parsedArgs.serials == null || parsedArgs.serials.isEmpty()) {
       builder.useAllAttachedDevices();
     } else {
       for (String serial : parsedArgs.serials) {
         builder.addDevice(serial);
+      }
+    }
+
+    //checks if there are any devices to skip testing on
+    if (parsedArgs.skipDevices != null && !parsedArgs.skipDevices.isEmpty()) {
+      for (String serial : parsedArgs.skipDevices) {
+        builder.skipDevice(serial);
       }
     }