+++ spoon-runner/src/main/java/com/squareup/spoon/html/HtmlIndex.java
@@ -8,10 +8,8 @@ import com.squareup.spoon.DeviceTestResult;
 import com.squareup.spoon.SpoonSummary;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import static com.squareup.spoon.DeviceTestResult.Status;
 
@@ -20,14 +18,12 @@ final class HtmlIndex {
   static HtmlIndex from(SpoonSummary summary) {
     int testsRun = 0;
     int totalSuccess = 0;
-    Set<DeviceTest> tests = new LinkedHashSet<DeviceTest>();
     List<Device> devices = new ArrayList<Device>();
     for (Map.Entry<String, DeviceResult> result : summary.getResults().entrySet()) {
       devices.add(Device.from(result.getKey(), result.getValue()));
       Map<DeviceTest, DeviceTestResult> testResults = result.getValue().getTestResults();
       testsRun += testResults.size();
       for (Map.Entry<DeviceTest, DeviceTestResult> entry : testResults.entrySet()) {
-        tests.add(entry.getKey());
         if (entry.getValue().getStatus() == Status.PASS) {
           totalSuccess += 1;
         }
@@ -57,18 +53,16 @@ final class HtmlIndex {
     }
     subtitle.append(" at ").append(started);
 
-    return new HtmlIndex(summary.getTitle(), subtitle.toString(), tests.size(), devices);
+    return new HtmlIndex(summary.getTitle(), subtitle.toString(),  devices);
   }
 
   public final String title;
   public final String subtitle;
-  public final int testCount;
   public final List<Device> devices;
 
-  HtmlIndex(String title, String subtitle, int testCount, List<Device> devices) {
+  HtmlIndex(String title, String subtitle, List<Device> devices) {
     this.title = title;
     this.subtitle = subtitle;
-    this.testCount = testCount;
     this.devices = devices;
   }
 
@@ -88,11 +82,13 @@ final class HtmlIndex {
     public final String name;
     public final List<TestResult> testResults;
     public final boolean executionFailed;
+    public final int testCount;
 
     Device(String serial, String name, List<TestResult> testResults, boolean executionFailed) {
       this.serial = serial;
       this.name = name;
       this.testResults = testResults;
+      this.testCount = testResults.size();
       this.executionFailed = executionFailed;
     }